fix(task): allow `imgUrl` to be empty string
Empty strings were interpreted as null, resulting in default URL. An
empty string is now valid, resulting in URL with only the image file,
which is a necessary use case that may not be covered by
'./filename.jpg'.

For example, Shopify allows liquid-based SCSS templates. One of their
liquid features is to dynamically generate a URL string based on a
filename, which itself is a key to an asset in a CDN:

    {{ 'filename.jpg' | asset_url }}

By allowing an empty imgUrl in this task, we can generate image
variables with only the filename, which in turn allows us to generate
code for use with said liquid feature. HGPA